TODO: Music override is bugged
If the previous song has looped, the accumulated mixer play time is longer than the song which produces an error when stopping music override.

Possible fix is to add an end event to music mixer and update the last player position when looping... We can possibly increment the loop count at this point as well so we can always restore the correct music state.

We should use milliseconds to compute these to minimize rounding errors.